Medical Bag bugfix
The Medical Bag did not handle that wounds or infections could be healed
by other sources (e.g. passive regeneration from Potion of Vitality)
while it was being treated - which typically lead to a crash.

Now the Medical Bag aborts the action (no supplies wasted) if it detects
that the treated status no longer exists.

+++ src/map_travel.cpp
@@ -96,7 +96,7 @@ void init()
 
 void save()
 {
-        saving::put_int(s_map_list.size());
+        saving::put_int((int)s_map_list.size());
 
         for (const MapType type : s_map_list)
         {
